Vastly improve the cost to store and copy Operator
Operator started as a small structure (a String and two words). But it
grew as we added fixity support, and now (according to Clippy) it's
over 128 bytes in size. Since it started small, I sloppy-copy it all
over the place, store Operators in error types, etc. So to get the
nice cheap copies and stores back, I put the operator implementation
into an Arc (sadly, it has to be thread-safe for Tauri). Now Operator
is incredibly cheap to copy and only takes up one pointer on the stack.