Manipulate borrow order in UnaryFunctionCommand
I want to pass &ApplicationState to UnaryFunctionCommand's function,
but to do that, I need to release the borrow on the stack first. It's
not less efficient, but it does make the code read a bit clunkier,
since we can't just do one broad stack borrow for the whole function.

I'll probably do the same for BinaryFunctionCommand eventually, but we
don't need it right now.