Remove TermParser abstraction, now that it's unnecessary (closes #134)
Since we decided to use two different multiplication operators rather
than having an overloaded scalar/matrix mode as Emacs does (see the
resolution on #87 for details on this implementation), the TermParser
is no longer necessary. The original plan for TermParser was to
parameterize it by the multiplication mode, but now we can implement
the parsing functions (which take no additional parameters, aside from
the expression to parse) directly on the Term type as associated functions.